+++ docs/Changelog.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Beta Updates
+* Add default download manager to download all files
+
+## v1.5.1
 * Release day is here!
 * Add full support for messaging in overlays. We will dynamically launch new overlays when required to.
 * Prevent bad messenger intent from launching
@@ -8,6 +11,7 @@
 * Add contextual menu items. Easily go to your full list of notifications or messages from the overlay.
 * Ensure that bottom bar layout does not hide the web content
 * Add option to share external links to Frost
+* Trigger notification service on each app start
 
 ## v1.4.13
 * Prevent image loading from trimming too many characters